(fix) Context menu on mobile preventing further swipe moves
When viewing the slider on mobile and performing a long press on the image opens the context menu, once you exit the menu endtouchHandler() is never called. You are no longer able to swipe left or right because touch.count is equal to 2 and touch.multitouch is equal to true.